Add Dependabot support for GitHub Actions
As part of setting up container building, we will be using a GitHub
Actions workflow to build container images.

It is good to keep our GitHub Actions workflow up-to-date.  So, have
Dependabot check our workflow files every week, and let us know if there
are any updates to the GitHub Actions that we use.