Add a script to be run by Docker
Before making the Dockerfile, it would be good to have a script that
does the work (the work of running the Perl scripts) for us.  So, here
is a script which does that!

The script first checks that the input directory is *not* empty.  If the
input directory is empty, the script fails.

NOTE: The HATS scripts only care about an input file whose name starts
with `hla_prot.fasta.`, and which end in a database version number.
Also, the HATS scripts might only support having a single input file.
This script **does not perform those additional checks**.

The script also checks that the output directory *is* empty.  If the
output directory is not empty, then this script fails.

After those checks, this script runs all of the HATS scripts, one at a
time.  Finally, a completion message is printed.

Fun Fact: This script could probably be used outside of Docker.  But,
the example commands are all `docker run` commands, so that's why this
is a Docker-specific script.