Add a basic Dockerfile
This is a Dockerfile to create a HATS container.

It starts with Rocky Linux 8 as a base, it upgrades packages (to the
latest-available 8.x), and it installs Perl.  The files from the Git
repo are copied in, then the `.git` directory & demo input file are
removed.  The working directory is set, and the dockerscript is set as
the default command.